🍎 Room Dao Delete Query 반환값 수정
## 🍎 Room Dao > Delete Query 함수 반환값 수정
- Long -> Int 수정
  > - DELETE 쿼리 메서드는 void 또는 int(삭제된 행 수)를 반환해야 함
  > **_[Stack Over Flow](https://stackoverflow.com/questions/70104556/android-room-kotlin-throws-delete-query-error)_**

+++ data/src/main/java/com/example/data/db/dao/MovieDao.kt
@@ -24,5 +24,5 @@ interface MovieDao {
     suspend fun deleteAll()
 
     @Delete
-    suspend fun deleteMovie(movieEntity: MovieEntity): Long
+    suspend fun deleteMovie(movieEntity: MovieEntity): Int
 }
